requirements.txt
streamlit==1.29.0
streamlit-authenticator==0.3.2
audio-recorder-streamlit==0.0.9
streamlit-float==0.3.5
google-api-python-client==2.132.0
google-auth==2.30.0
google_auth_oauthlib==1.2.0
openai==1.31.1